Data Science pratice
NOTE: This users a helper python file that I thought was helpful so I stripped it down and used it as part of this project. It basically uses sklearn and wraps it up so you can get more metrics and things since sklearn doesn't provide that out of the box.